caprevoke: Introduce VM_CAPREVOKE_PMAP_SYNC
Explicitly signal vm_caprevoke with whether we desire a synchronization
with the PMAP's capdirty bits.  Experimentation has suggested that the
PMAP may be a large cache of dirty bits that get flushed back to us, and
so it is likely worth doing another IPI even on concurrent passes; we
shall see.

Expose CAPREVOKE_EARLY_NO_SYNC to userland to control this behaviour;
asserting it will revert the system to the current behavior of doing
PMAP synchronization only once the world is stopped.