caprevoke: handle pages from shadowed objects better
Apparently it is possible for vm_fault_hold(VM_PROT_READ) to return
a page which is writable according to pmap_page_is_write_mapped and
yet for which its containing object has refcount greater than one.
Mutating this page seems like a mistake, so force it through the usual
RO-upgrade-to-RW logic instead.