caprevoke machdep: remove onfault set/reset from inner loop
This is abjectly terrifying: directly dereferencing "user" memory.  It's
safe because the shadow maps are fully managed by the kernel and not
subject to unmapping by userland.  Nevertheless, we take TLB miss (and
eventually, modify) faults in the kernel; placate the invariants of
MIPS' trap.c by registering a ->onfault "handler" that panics the kernel
if it's ever called.  Unlike the earlier code, which did this dance once
per bitmap byte read, we just set and reset ->onfault once per page.