caprevoke machdep: change page iterator interface
Again, make it a little easier for the optimizer to see what's going on.
Concretely, rather than using a bit in the scanned state
(VM_CAPREVOKE_PAGE_EARLY_OUT) to signal early return from the iterator,
have the callback *return* the early-out signal and shift the other
result bits (VM_CAPREVOKE_PAGE_*) to an out-parameter.

The optimizer can directly see the conditional control transfers and
eliminate the materialized return value and simultaneously hoists the
result out parameter into a register, as desired.