caprevoke machdep: read cloadtags_stride once per page
Rather than reloading every iteration; we can't declare it constant
because we need to update it once at bootup.  This is almost as good,
since we don't expect the optimizer to be able to hoist the reads
higher, as the caller is in a different file and is a fairly complex
beast.