caprevoke: put the tag checks back
cgetbase can yield an "address" which does not map to the shadow region
if the loaded capability is untagged.  This then panics the kernel when
we go to read the "shadow" bit, because, well, it isn't there.

While we'd really like to have the tag tests be on the slightly colder
path of being perceived as needing revocation, we can't do that.